Displaying the newly added note first
Sorts the notes in descending order of their dates created. For this, I had to include a new key for the database called 'keyForSort' with datatype of Date. It stores the current system time of the user when he/she creates the note. This is then used in backend- note.js to sort the Note.find() method in descending order of their dates created.